refactor(remote): extract remote_support.py to keep config under the size gate
- ouroboros/remote_support.py: get/update_remote_connections + server pid-lock
  helpers moved out of config (config crossed the 1600-line smoke gate; P7
  says extract, not grandfather); config re-exports them so config.X stays the
  stable API and lazy config access inside remote_support avoids an import cycle
- tests updated: server-lock handle now lives in remote_support; files_ui
  bridge test asserts the new active-view-port validation contract
